# Sema: rewrite semantic analysis of function calls