--- a/lib/std/math/log_int.zig
+++ b/lib/std/math/log_int.zig
@@ -61,6 +61,7 @@ pub fn log_int(comptime T: type, base: T, x: T) Log2Int(T) {
 }
 
 test "log_int" {
+    @setEvalBranchQuota(2000);
     // Test all unsigned integers with 2, 3, ..., 64 bits.
     // We cannot test 0 or 1 bits since base must be > 1.
     inline for (2..64 + 1) |bits| {