--- a/test/cases/compile_errors/generic_function_instantiation_inherits_parent_branch_quota.zig
+++ b/test/cases/compile_errors/generic_function_instantiation_inherits_parent_branch_quota.zig
@@ -22,9 +22,8 @@ fn Type(comptime n: usize) type {
 }
 
 // error
-// backend=stage2
-// target=native
 //
 // :21:16: error: evaluation exceeded 1001 backwards branches
 // :21:16: note: use @setEvalBranchQuota() to raise the branch limit from 1001
 // :16:34: note: called from here
+// :8:15: note: called from here