--- a/test/cases/compile_errors/inline_call_runtime_value_to_comptime_param.zig
+++ b/test/cases/compile_errors/inline_call_runtime_value_to_comptime_param.zig
@@ -10,8 +10,7 @@ pub export fn entry() void {
 }
 
 // error
-// backend=stage2
-// target=native
 //
 // :5:18: error: unable to resolve comptime value
 // :5:18: note: argument to comptime parameter must be comptime-known
+// :1:24: note: parameter declared comptime here