--- a/test/cases/compile_errors/invalid_extern_function_call.zig
+++ b/test/cases/compile_errors/invalid_extern_function_call.zig
@@ -9,8 +9,6 @@ export fn entry1() void {
 }
 
 // error
-// backend=stage2
-// target=native
 //
-// :4:15: error: comptime call of extern function pointer
-// :8:5: error: inline call of extern function pointer
+// :4:15: error: comptime call of extern function
+// :8:5: error: inline call of extern function