This rewrite improves some error messages, hugely simplifies the logic, and fixes several bugs. One of these bugs is technically a new rule which Andrew and I agreed on: if a parameter has a comptime-only type but is not declared `comptime`, then the corresponding call argument should not be *evaluated* at comptime; only resolved. Implementing this required changing how function types work a little, which in turn required allowing a new kind of function coercion for some generic use cases: function coercions are now allowed to implicitly *remove* `comptime` annotations from parameters with comptime-only types. This is okay because removing the annotation affects only the call site.

Resolves: #22262

This **is** a breaking change, because it fixes a bug relating to eval branch quota. Previously, the usage of a "child Sema" when analyzing a generic function's call site meant we accidentally "duplicated" the eval branch quota. We no longer do this (I've actually concluded that the concept of a "child Sema" isn't really coherent), so a little more eval branch quota is used up on evaluating generic parameter/return types. That's why the changes to `lib/compiler/aro_translate_c.zig` and `src/translate_c.zig` were necessary. #22410 is **not** resolved by this PR; the buggy behavior of status quo is left intact.

I was considering also reworking the ZIR for function declarations in this PR, since it's very overcomplicated. However, I decided to leave that for now, since I'm planning to start playing around with a biiiig ZIR rewrite soon.