@mlugg 2025-01-09
## Performance Data Points

### Analyze `std` Tests

```
Benchmark 1 (6 runs): /home/mlugg/zig/master/stage4-fast/bin/zig test --zig-lib-dir lib lib/std/std.zig -fno-emit-bin
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          10.1s  ± 12.3ms    10.1s  … 10.1s           0 ( 0%)        0%
  peak_rss            595MB ±  170KB     595MB …  595MB          0 ( 0%)        0%
  cpu_cycles         43.2G  ± 39.5M     43.2G  … 43.3G           0 ( 0%)        0%
  instructions       84.6G  ± 37.2K     84.6G  … 84.6G           0 ( 0%)        0%
  cache_references   3.21G  ± 8.74M     3.21G  … 3.23G           0 ( 0%)        0%
  cache_misses       80.4M  ± 2.08M     78.1M  … 84.0M           0 ( 0%)        0%
  branch_misses      63.8M  ±  192K     63.5M  … 64.0M           0 ( 0%)        0%
Benchmark 2 (7 runs): /home/mlugg/zig/better-analyze-call/stage4-fast/bin/zig test --zig-lib-dir lib lib/std/std.zig -fno-emit-bin
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          9.97s  ± 20.4ms    9.94s  … 9.99s           0 ( 0%)        ⚡-  1.6% ±  0.2%
  peak_rss            594MB ±  201KB     594MB …  594MB          0 ( 0%)          -  0.2% ±  0.0%
  cpu_cycles         42.5G  ± 90.4M     42.4G  … 42.7G           0 ( 0%)        ⚡-  1.6% ±  0.2%
  instructions       83.5G  ± 21.5K     83.5G  … 83.5G           0 ( 0%)        ⚡-  1.3% ±  0.0%
  cache_references   3.10G  ± 8.62M     3.09G  … 3.12G           0 ( 0%)        ⚡-  3.6% ±  0.3%
  cache_misses       84.0M  ± 2.06M     81.3M  … 86.4M           0 ( 0%)        💩+  4.5% ±  3.2%
  branch_misses      63.9M  ±  645K     63.0M  … 64.9M           0 ( 0%)          +  0.2% ±  0.9%
```

### Analyze Compiler Tests

```
Benchmark 1 (8 runs): /home/mlugg/zig/master/stage4-fast/bin/zig build-exe --dep aro --dep aro_translate_c --dep build_options -Mroot=src/main.zig -Maro=lib/compiler/aro/aro.zig --dep aro -Maro_translate_c=lib/compiler/aro_translate_c.zig -Mbuild_options=.zig-cache/c/9883e0613966ebf6eeefe29d780bb845/options.zig -fno-emit-bin --zig-lib-dir lib
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          7.58s  ± 56.5ms    7.53s  … 7.72s           2 (25%)        0%
  peak_rss            389MB ±  150KB     389MB …  389MB          0 ( 0%)        0%
  cpu_cycles         32.5G  ±  222M     32.3G  … 33.0G           1 (13%)        0%
  instructions       58.8G  ± 22.1K     58.8G  … 58.8G           0 ( 0%)        0%
  cache_references   2.78G  ± 9.54M     2.76G  … 2.79G           0 ( 0%)        0%
  cache_misses        126M  ± 1.12M      124M  …  127M           0 ( 0%)        0%
  branch_misses      89.8M  ±  546K     89.3M  … 91.0M           1 (13%)        0%
Benchmark 2 (9 runs): /home/mlugg/zig/better-analyze-call/stage4-fast/bin/zig build-exe --dep aro --dep aro_translate_c --dep build_options -Mroot=src/main.zig -Maro=lib/compiler/aro/aro.zig --dep aro -Maro_translate_c=lib/compiler/aro_translate_c.zig -Mbuild_options=.zig-cache/c/9883e0613966ebf6eeefe29d780bb845/options.zig -fno-emit-bin --zig-lib-dir lib
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          7.45s  ± 50.8ms    7.40s  … 7.54s           0 ( 0%)        ⚡-  1.8% ±  0.7%
  peak_rss            384MB ±  193KB     384MB …  384MB          0 ( 0%)        ⚡-  1.3% ±  0.0%
  cpu_cycles         31.9G  ±  210M     31.7G  … 32.3G           0 ( 0%)        ⚡-  1.7% ±  0.7%
  instructions       57.5G  ± 24.4K     57.5G  … 57.5G           0 ( 0%)        ⚡-  2.3% ±  0.0%
  cache_references   2.66G  ± 11.3M     2.65G  … 2.69G           0 ( 0%)        ⚡-  4.1% ±  0.4%
  cache_misses        127M  ± 2.31M      124M  …  131M           0 ( 0%)          +  1.2% ±  1.5%
  branch_misses      88.5M  ±  552K     87.9M  … 89.5M           0 ( 0%)          -  1.4% ±  0.6%
```

TL;DR: faster :sunglasses:

## Maximum Comptime Recursion Depth

Running `zig build-obj foo.zig` on this file:

```zig
comptime {
    @setEvalBranchQuota(10_000);
    @compileLog(fib(1234)); // replacing this value
}

fn fib(n: comptime_int) comptime_int {
    if (n == 0) return 1;
    return fib(n - 1) * n;
}
```

`master`: 4062
This branch: 2854 (-30%)

TL;DR: worse :cry:

@andrewrk, if this is an issue, I can probably improve it by splitting out the argument-evaluating logic (which has a `Block` and stuff) into a separate function to decrease the size of the main inline call stack frame.