@dnjulek 2025-01-11
It seems that this has broken ZLS build

```
zls git:(master) zig build -Doptimize=ReleaseSafe
install
└─ install zls
   └─ zig build-exe zls ReleaseSafe native 1 errors
/usr/local/lib/zig/std/mem.zig:669:5: error: evaluation exceeded 1000 backwards branches
    for (a, b) |a_elem, b_elem| {
    ^~~
/usr/local/lib/zig/std/mem.zig:669:5: note: use @setEvalBranchQuota() to raise the branch limit from 1000
.zig-cache/o/62617b606a626867bfe5e239ceb48dc2/lsp_types.zig:41:24: note: called from here
        if (std.mem.eql(u8, method, meta.method)) {
            ~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~
```
          
zig: 0.14.0-dev.2634+b36ea592b
zls: a9e651a2456893e41c2a72af354a226a953a8253