@mlugg 2025-01-11
ZLS is unaffiliated with the Zig project -- please open an issue on the ZLS repo instead.